checkpoint

+++ Box2D/Box2D/Dynamics/Contacts/b2Contact.h
@@ -243,11 +243,11 @@ inline void b2Contact::SetEnabled(bool flag)
 {
 	if (flag)
 	{
-		m_flags |= e_enabledFlag;
+		m_flags |= static_cast<unsigned int>(e_enabledFlag);
 	}
 	else
 	{
-		m_flags &= ~e_enabledFlag;
+		m_flags &= ~(static_cast<unsigned int>(e_enabledFlag));
 	}
 }
